requirements.txt
django
djangorestframework
djangorestframework-simplejwt
django-cors-headers
customlib/dist/helperlibrary-1.0.tar.gz